Bugfixes and improvements

+++ src/core/defaults/.boilerplate/api/apiDataDelete.js
@@ -17,36 +17,42 @@ export default () => ({
                     message: "validation_error"
                 });
             }
-            const formData = new FormData();
-            const query = {
-                $and: [{
-                    $or: [],
-                }, {
-                    ...(utils.filter({}, authData) || {}),
-                }],
-            };
-            for (const id of req.body.ids) {
-                query.$and[0].$or.push({
-                    _id: new ObjectId(id)
-                });
-            }
-            if (formData.getRecycleBinConfig && formData.getRecycleBinConfig().enabled) {
-                const updateResult = await this.mongo.db.collection(moduleConfig.collections.main).updateMany(query, {
-                    $set: {
-                        deleted: {
-                            date: new Date(),
-                            userId: authData._id.toString(),
-                        },
-                    }
-                });
+            if (!this.systemConfig.demo) {
+                const formData = new FormData();
+                const query = {
+                    $and: [{
+                        $or: [],
+                    }, {
+                        ...(utils.filter({}, authData) || {}),
+                    }],
+                };
+                for (const id of req.body.ids) {
+                    query.$and[0].$or.push({
+                        _id: new ObjectId(id)
+                    });
+                }
+                if (formData.getRecycleBinConfig && formData.getRecycleBinConfig().enabled) {
+                    const updateResult = await this.mongo.db.collection(moduleConfig.collections.main).updateMany(query, {
+                        $set: {
+                            deleted: {
+                                date: new Date(),
+                                userId: authData._id.toString(),
+                            },
+                        }
+                    });
+                    return rep.code(200).send({
+                        count: updateResult.modifiedCount,
+                    });
+                }
+                const deleteResult = await this.mongo.db.collection(moduleConfig.collections.main).deleteMany(query);
                 return rep.code(200).send({
-                    count: updateResult.modifiedCount,
+                    count: deleteResult.deletedCount,
+                });
+            } else {
+                return rep.code(200).send({
+                    count: 0,
                 });
             }
-            const deleteResult = await this.mongo.db.collection(moduleConfig.collections.main).deleteMany(query);
-            return rep.code(200).send({
-                count: deleteResult.deletedCount,
-            });
         } catch (e) {
             this.log.error(e);
             return Promise.reject(e);
