+++ src/core/defaults/.boilerplate/api/apiHistoryDelete.js
@@ -38,9 +38,14 @@ export default () => ({
                     message: "notFound"
                 }, 404);
             }
-            const deleteResult = await this.mongo.db.collection(this.systemConfig.collections.history).deleteMany(query);
+            if (!this.systemConfig.demo) {
+                const deleteResult = await this.mongo.db.collection(this.systemConfig.collections.history).deleteMany(query);
+                return rep.code(200).send({
+                    count: deleteResult.deletedCount,
+                });
+            }
             return rep.code(200).send({
-                count: deleteResult.deletedCount,
+                count: 0,
             });
         } catch (e) {
             this.log.error(e);
